requirements.txt
python-telegram-bot==20.8
openai==1.3.0
chromadb==0.4.22
python-dotenv==1.0.0
PyPDF2==3.0.1
python-multipart==0.0.6
fastapi==0.109.0
uvicorn==0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
langchain==0.1.0
tiktoken==0.5.2
